Backend unit testing

+++ backend/routes/buyer.js
@@ -40,24 +40,7 @@ router.post('/addCartRegItem/:id', async (req, res) => {
       { $addToSet: { shoppingCart: itemId }});
     res.status(200).send('Regular listing successfully added to cart!');
   } catch (error) {
-    res.status(500).send('An unknown error occured');
-    throw new Error('Error with adding reg item to cart');
-  }
-});
-
-// route to add bid item to cart
-router.post('/addCartBidItem/:id', async (req, res) => {
-  try {
-    const { bid } = req.body;
-    const item = await ItemBid.findById(req.params.id);
-    await User.findOneAndUpdate(
-      { email: req.session.email,
-        'shoppingCart._id': { $ne: req.params.id }},
-      { $addToSet: { shoppingCart: {item , bid} }});
-    res.status(200).send('Regular listing successfully added to cart!');
-  } catch (error) {
-    res.status(500).send('An unknown error occured');
-    throw new Error('Error with adding bid item to cart');
+    res.status(500).send('Error adding item to cart');
   }
 });
 
@@ -71,8 +54,7 @@ router.post('/removeCartRegItem/:id', async (req, res) => {
       { $pull: { shoppingCart: itemId }});
     res.status(200).send('Regular listing removed successfully from cart!');
   } catch (error) {
-    res.status(500).send('An unknown error occured');
-    throw new Error('Error with removing reg item from cart');
+    res.status(500).send('Error removing item from cart');
   }
 });
 
@@ -97,8 +79,7 @@ router.post('/addWatchRegItem/:id', async (req, res) => {
       { $addToSet: { watchlistRegular: item }});
     res.status(200).send('Regular listing successfully added to watchlist!');
   } catch (error) {
-    res.status(500).send('An unknown error occured');
-    throw new Error('Error with adding reg item to watchlist');
+    res.status(500).send('Error with adding reg item to watchlist');
   }
 });
 
@@ -112,11 +93,11 @@ router.post('/addWatchBidItem/:id', async (req, res) => {
       { $addToSet: { watchlistBid: item }});
     res.status(200).send('Bid listing successfully added to watchlist!');
   } catch (error) {
-    res.status(500).send('An unknown error occured');
-    throw new Error('Error with adding bid item to watchlist');
+    res.status(500).send('Error with adding bid item to watchlist');
   }
 });
 
+/*
 // route to remove reg item from watchlist
 router.post('/removeWatchRegItem/:id', async (req, res) => {
   try {
@@ -132,7 +113,7 @@ router.post('/removeWatchRegItem/:id', async (req, res) => {
 });
 
 // route to remove bid item from watchlist
-router.post('/removeWatchRegItem/:id', async (req, res) => {
+router.post('/removeWatchBidItem/:id', async (req, res) => {
   try {
     const item = await ItemBid.findById(req.params.id);
     await User.findOneAndUpdate(
@@ -144,6 +125,7 @@ router.post('/removeWatchRegItem/:id', async (req, res) => {
     throw new Error('Error with removing bid item from watchlist');
   }
 });
+*/
 
 // route to add bid to bidhistory
 router.post('/addBid/:id', async (req, res) => {
@@ -157,8 +139,7 @@ router.post('/addBid/:id', async (req, res) => {
     );
     res.status(200).send('Bid placed successfully');
   } catch (error) {
-    res.status(500).send('An unknown error occured');
-    throw new Error('Error with adding bid to item');
+    res.status(500).send('Error with adding bid to item');
   }
 });
 
@@ -179,7 +160,7 @@ router.post('/regTransaction', async (req, res) => {
     });
     res.status(201).json(transaction);
   } catch (error) {
-    throw new Error('Error with completing transaction');
+    res.status(500).send('Error with completing Transaction');
   }
 });
 
@@ -217,10 +198,12 @@ router.post('/addTransaction', async (req, res) => {
       { name: transaction.seller },
       { $addToSet: { transactionHistory: transaction } },
     );
-    await ItemRegular.findByIdAndDelete(transaction.listingRegular._id);
-    res.status(200).send('Regular listing successfully added to watchlist!');
+    if (transaction.listingRegular) {
+      await ItemRegular.findByIdAndDelete(transaction.listingRegular._id);
+    }
+    res.status(200).send('Transaction successfully processed');
   } catch (error) {
-    throw new Error('Error with completeing transaction');
+    res.status(500).send('Error with completing Transaction');
   }
 });
